+++ python/GafferUSDUI/__init__.py
@@ -38,5 +38,6 @@ from . import USDAttributesUI
 from . import USDLayerWriterUI
 from . import USDShaderUI
 from . import USDLightUI
+from . import _PointInstancerAdaptorUI
 
 __import__( "IECore" ).loadConfig( "GAFFER_STARTUP_PATHS", subdirectory = "GafferUSDUI" )